[#39] 로그인된 유저 정보 조회 API 이동
auth 모듈에서 members 모듈로 이동함.
