+++ src/members/dtos/get-member.dto.ts
@@ -1,4 +1,5 @@
 import { ApiProperty, OmitType } from '@nestjs/swagger';
+import { IsOptional } from 'class-validator';
 import { CoreOutput } from 'src/common/dtos/output.dto';
 import { Member } from '../entities/members.entity';
 
@@ -6,4 +7,11 @@ export class GetMemberEntity extends OmitType(Member, ['mbrKakaoSeq']) {}
 export class GetMemberOutput extends CoreOutput {
     @ApiProperty()
     member?: GetMemberEntity;
+}
+
+export class getLoggedInMemberOutput extends CoreOutput {
+    @IsOptional()
+    response?: {
+        member: Member;
+    };
 }
